Woohoo! First refactor into a class
